requirements.txt
Flask
python-dotenv
python-whois
dnspython
requests
pyinstaller
Pillow
